Cleaned up a lot of stuff (some worthy code was temporarily removed)
* The worthy code will be re-introduced in the minor releases
* Reviewed lots of "FIX" todos, fixed some small stuff
* Some tests adjustments
* Pre-Final reviews/refactorings before release

+++ tests/general/PHPHelperTest.php
@@ -356,7 +356,7 @@ class PHPHelperTest extends TestCase {
 			['this is string', 'string'],
 			['anotherstringishere', 'string'],
 			[12, 'integer'],
-			[22.22, 'double'],
+			[22.22, 'float'],
 			[new $version_class('0.0.0', 'no app'), $version_class],
 			[PHP::info(), $phpinfo_class],
 			[true, 'boolean'],
